+++ src/db/seeders/seeds-down.ts
@@ -1,0 +1,15 @@
+import { UserSeeds } from './user.seeds';
+import { ProjectSeeds } from './project.seeds';
+
+import { sequelize } from '../models';
+
+const SeedUp = async () => {
+  await sequelize().sync();
+
+  await UserSeeds.down();
+  await ProjectSeeds.down();
+
+  return;
+};
+
+SeedUp();