+++ src/errors/log-error.ts
@@ -1,0 +1,5 @@
+import { logger } from '../server';
+
+export const logError = (error: Error, message: string) => {
+  logger.error(message, JSON.stringify(error));
+};