+++ src/errors/throw-error.ts
@@ -1,0 +1,6 @@
+import { logError } from './log-error';
+
+export const throwError = (error: Error, message: string = 'OOPS!') => {
+  logError(error, message);
+  throw new Error(message);
+};