+++ src/mongoose/client.ts
@@ -1,18 +1,19 @@
 import * as mongoose from 'mongoose';
-import { MONGO_URI } from '../../constants';
+import { MONGO_URI } from '../constants';
+import { logger } from '../server';
 
-export const db = async () => {
+export const mongo = async () => {
   if (process.env.NODE_ENV !== 'test') {
     return mongoose.connect(
       `${MONGO_URI}`,
       { useNewUrlParser: true, useUnifiedTopology: true },
       (err) => {
         if (err) {
-          console.log(
+          logger.error(
             `MongoDB connection error. Please make sure MongoDB is running. Error:${err}`
           );
         } else {
-          console.log('****\n Connected to MongoDB! \n****');
+          logger.info('****\n Connected to MongoDB! \n****');
         }
       }
     );