+++ src/spec/spec-helpers.ts
@@ -1,4 +1,12 @@
 import * as supertest from 'supertest';
-import { app } from '../server';
+import { app, redis } from '../server';
+import { models } from '../db/models';
 
+// Express
 export const request = supertest(app);
+
+// Cleanup Tests
+afterAll(() => {
+  models.map(async (m: any) => m.sequelize?.close());
+  redis.quit();
+});